requirements.txt
torch
numpy
tqdm
nltk
scikit-learn
torchvision
flask
flask-ngrok
pillow
tkinter
nltk.translate
albumentations 
albumentations[pytorch]
nltk
spacy
Pillow
torchvision
scikit-learn
